Update the project to use Gradle
- updated to the 1.13 PaperSpigot API
- due to the fact above, changed a few material names (untested,
  sorry <3)
- added a .gitignore
- remove a bunch of stuff that shouldn't be in the repo
