+++ README.md
@@ -1,0 +1,15 @@
+## ExpressionsPlus
+
+#### Compiling
+
+Windows:
+```bash
+gradlew jar
+```
+
+Linux/Mac:
+```bash
+./gradlew jar
+```
+
+or use your favorite IDE to execute the `jar` Gradle task
