+++ src/main/java/de/jaylawl/expressionsplus/elements/expressions/ExprMaxFuseTicks.java
