+++ src/main/java/de/jaylawl/expressionsplus/elements/expressions/ExprPoweredState.java
