+++ src/main/java/de/jaylawl/expressionsplus/elements/expressions/ExprRiches.java
