+++ src/main/java/de/jaylawl/expressionsplus/elements/expressions/ExprSpawnRange.java
@@ -1,40 +1,36 @@
 package de.jaylawl.expressionsplus.elements.expressions;
 
 import ch.njol.skript.classes.Changer.ChangeMode;
-import ch.njol.skript.entity.EntityData;
 import ch.njol.skript.expressions.base.SimplePropertyExpression;
 import ch.njol.util.coll.CollectionUtils;
-import ch.njol.skript.entity.EntityType;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockState;
 import org.bukkit.block.CreatureSpawner;
-import org.bukkit.entity.Entity;
 import org.bukkit.event.Event;
 
 import javax.annotation.Nullable;
 
-public class ExprSpawnedType extends SimplePropertyExpression<Block, EntityType> {
+public class ExprSpawnRange extends SimplePropertyExpression<Block, Number> {
 
     static {
-        register(ExprSpawnedType.class, EntityType.class, "spawned ([entity ]type|type[ of entity])", "block");
+        register(ExprSpawnRange.class, Number.class, "[entity( |-)]spawn range", "block");
     }
 
     @Override
     @Nullable
-    public EntityType convert(Block block) {
-        if (block.getType() != Material.MOB_SPAWNER)
+    public Number convert(Block block) {
+        if (block.getType() != Material.SPAWNER)
             return null;
         BlockState state = block.getState();
         CreatureSpawner spawner = (CreatureSpawner) state;
-        return EntityType.parse(spawner.getSpawnedType().toString().replace("_", " "));
+        return spawner.getSpawnRange();
     }
 
-    /*
     @Override
     public Class<?>[] acceptChange(final ChangeMode mode) {
-        if (mode == ChangeMode.SET)
-            return CollectionUtils.array(EntityType.class);
+        if (mode == ChangeMode.ADD || mode == ChangeMode.REMOVE || mode == ChangeMode.SET || mode == ChangeMode.RESET)
+            return CollectionUtils.array(Number.class);
         return null;
     }
 
@@ -42,13 +38,22 @@ public class ExprSpawnedType extends SimplePropertyExpression<Block, EntityType>
     public void change(Event event, Object[] delta, ChangeMode mode){
         if (delta != null) {
             Block block = getExpr().getSingle(event);
-            if (block.getType() == Material.MOB_SPAWNER) {
+            if (block.getType() == Material.SPAWNER) {
                 BlockState state = block.getState();
                 CreatureSpawner spawner = (CreatureSpawner) state;
-                EntityType value = (EntityType) delta[0];
+                Integer value = ((Number) delta[0]).intValue();
                 switch (mode) {
+                    case ADD:
+                        spawner.setSpawnRange(spawner.getSpawnRange() + value);
+                        break;
+                    case REMOVE:
+                        spawner.setSpawnRange(spawner.getSpawnRange() - value);
+                        break;
                     case SET:
-                        spawner.setSpawnedType(val2);
+                        spawner.setSpawnRange(value);
+                        break;
+                    case RESET:
+                        spawner.setSpawnRange(4);
                         break;
                     default:
                         assert false;
@@ -56,15 +61,14 @@ public class ExprSpawnedType extends SimplePropertyExpression<Block, EntityType>
             }
         }
     }
-    */
 
     @Override
     protected String getPropertyName() {
-        return "spawned entity type";
+        return "entity spawn range";
     }
 
     @Override
-    public Class<? extends EntityType> getReturnType() {
-        return EntityType.class;
+    public Class<? extends Number> getReturnType() {
+        return Number.class;
     }
 }
