+++ src/main/java/de/jaylawl/expressionsplus/elements/expressions/ExprSpawnedType.java
@@ -1,36 +1,40 @@
 package de.jaylawl.expressionsplus.elements.expressions;
 
 import ch.njol.skript.classes.Changer.ChangeMode;
+import ch.njol.skript.entity.EntityData;
 import ch.njol.skript.expressions.base.SimplePropertyExpression;
 import ch.njol.util.coll.CollectionUtils;
+import ch.njol.skript.entity.EntityType;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockState;
 import org.bukkit.block.CreatureSpawner;
+import org.bukkit.entity.Entity;
 import org.bukkit.event.Event;
 
 import javax.annotation.Nullable;
 
-public class ExprMaxNearbyEntities extends SimplePropertyExpression<Block, Number> {
+public class ExprSpawnedType extends SimplePropertyExpression<Block, EntityType> {
 
     static {
-        register(ExprMaxNearbyEntities.class, Number.class, "max[imum] (entities nearby|nearby entities)", "block");
+        register(ExprSpawnedType.class, EntityType.class, "spawned ([entity ]type|type[ of entity])", "block");
     }
 
     @Override
     @Nullable
-    public Number convert(Block block) {
-        if (block.getType() != Material.MOB_SPAWNER)
+    public EntityType convert(Block block) {
+        if (block.getType() != Material.SPAWNER)
             return null;
         BlockState state = block.getState();
         CreatureSpawner spawner = (CreatureSpawner) state;
-        return spawner.getMaxNearbyEntities();
+        return EntityType.parse(spawner.getSpawnedType().toString().replace("_", " "));
     }
 
+    /*
     @Override
     public Class<?>[] acceptChange(final ChangeMode mode) {
-        if (mode == ChangeMode.ADD || mode == ChangeMode.REMOVE || mode == ChangeMode.SET || mode == ChangeMode.RESET)
-            return CollectionUtils.array(Number.class);
+        if (mode == ChangeMode.SET)
+            return CollectionUtils.array(EntityType.class);
         return null;
     }
 
@@ -41,19 +45,10 @@ public class ExprMaxNearbyEntities extends SimplePropertyExpression<Block, Numbe
             if (block.getType() == Material.MOB_SPAWNER) {
                 BlockState state = block.getState();
                 CreatureSpawner spawner = (CreatureSpawner) state;
-                Integer value = ((Number) delta[0]).intValue();
+                EntityType value = (EntityType) delta[0];
                 switch (mode) {
-                    case ADD:
-                        spawner.setMaxNearbyEntities(spawner.getMaxNearbyEntities() + value);
-                        break;
-                    case REMOVE:
-                        spawner.setMaxNearbyEntities(spawner.getMaxNearbyEntities() - value);
-                        break;
                     case SET:
-                        spawner.setMaxNearbyEntities(value);
-                        break;
-                    case RESET:
-                        spawner.setMaxNearbyEntities(6);
+                        spawner.setSpawnedType(val2);
                         break;
                     default:
                         assert false;
@@ -61,14 +56,15 @@ public class ExprMaxNearbyEntities extends SimplePropertyExpression<Block, Numbe
             }
         }
     }
+    */
 
     @Override
     protected String getPropertyName() {
-        return "max nearby entities";
+        return "spawned entity type";
     }
 
     @Override
-    public Class<? extends Number> getReturnType() {
-        return Number.class;
+    public Class<? extends EntityType> getReturnType() {
+        return EntityType.class;
     }
 }
