+++ src/main/java/de/jaylawl/expressionsplus/elements/expressions/ExprTNTFuseTicks.java
