+++ src/main/java/de/jaylawl/expressionsplus/elements/expressions/ExprWolfAnger.java
